Enum refactoring. Remove redefinition
