+++ backend/dataall/modules/dataset_sharing/api/types.py
@@ -1,5 +1,5 @@
 from dataall.base.api import gql
-from dataall.modules.dataset_sharing.api.enums import ShareableType, PrincipalType
+from dataall.modules.dataset_sharing.common.enums import ShareableType, PrincipalType
 from dataall.modules.dataset_sharing.api.resolvers import union_resolver, resolve_shared_item, resolve_dataset, \
     resolve_consumption_data, resolve_existing_shared_items, resolve_share_object_statistics, resolve_principal, \
     resolve_group, list_shareable_objects, resolve_user_role
