+++ backend/dataall/modules/dataset_sharing/db/share_object_models.py
@@ -5,7 +5,7 @@ from sqlalchemy import Boolean, Column, String, DateTime
 from sqlalchemy.orm import query_expression
 
 from dataall.base.db import Base, utils
-from dataall.modules.dataset_sharing.db.enums import ShareObjectStatus, ShareItemStatus
+from ..common.enums import ShareObjectStatus, ShareItemStatus
 
 
 def in_one_month():
