+++ backend/dataall/modules/dataset_sharing/services/data_sharing_service.py
@@ -1,5 +1,8 @@
 import logging
 
+from dataall.base.db import Engine
+from dataall.modules.dataset_sharing.db.share_object_repositories import ShareObjectSM, ShareObjectRepository, \
+    ShareItemSM
 from dataall.modules.dataset_sharing.services.share_processors.lf_process_cross_account_share import \
     ProcessLFCrossAccountShare
 from dataall.modules.dataset_sharing.services.share_processors.lf_process_same_account_share import \
@@ -8,11 +11,7 @@ from dataall.modules.dataset_sharing.services.share_processors.s3_access_point_p
     ProcessS3AccessPointShare
 from dataall.modules.dataset_sharing.services.share_processors.s3_bucket_process_share import ProcessS3BucketShare
 
-from dataall.base.db import Engine
-from dataall.modules.dataset_sharing.db.enums import (ShareObjectActions, ShareItemStatus, ShareableType,
-                                                      ShareItemActions)
-from dataall.modules.dataset_sharing.db.share_object_repositories import ShareObjectSM, ShareObjectRepository, \
-    ShareItemSM
+from ..common.enums import (ShareObjectActions, ShareItemStatus, ShareableType)
 
 log = logging.getLogger(__name__)
