+++ backend/dataall/modules/datasets/api/dataset/input_types.py
@@ -1,6 +1,6 @@
 from dataall.base.api import gql
 from dataall.base.api.constants import SortDirection
-from dataall.modules.datasets.api.dataset.enums import DatasetSortField
+from dataall.modules.datasets_base.constants.enums import DatasetSortField
 
 
 NewDatasetInput = gql.InputType(
