+++ backend/dataall/modules/datasets/api/dataset/types.py
@@ -1,5 +1,5 @@
 from dataall.base.api import gql
-from dataall.modules.datasets.api.dataset.enums import DatasetRole
+from dataall.modules.datasets_base.constants.enums import DatasetRole
 from dataall.modules.datasets.api.dataset.resolvers import (
     get_dataset_environment,
     get_dataset_organization,
@@ -14,7 +14,6 @@ from dataall.modules.datasets.api.dataset.resolvers import (
     get_dataset_stack
 )
 from dataall.core.environment.api.enums import EnvironmentPermission
-
 
 DatasetStatistics = gql.ObjectType(
     name='DatasetStatistics',
@@ -153,7 +152,6 @@ Dataset = gql.ObjectType(
     ],
 )
 
-
 DatasetSearchResult = gql.ObjectType(
     name='DatasetSearchResult',
     fields=[
@@ -168,7 +166,6 @@ DatasetSearchResult = gql.ObjectType(
         gql.Field(name='hasPrevious', type=gql.Boolean),
     ],
 )
-
 
 GlueCrawler = gql.ObjectType(
     name='GlueCrawler',
