+++ backend/dataall/modules/datasets/api/table/enums.py
@@ -1,12 +1,6 @@
 from dataall.base.api.constants import GraphQLEnumMapper
 
 
-class DatasetSortField(GraphQLEnumMapper):
-    created = 'created'
-    updated = 'updated'
-    label = 'label'
-
-
 class DatasetTablePreviewStatus(GraphQLEnumMapper):
     QUEUED = 'QUEUED'
     RUNNING = 'RUNNING'
