+++ backend/dataall/modules/datasets_base/db/dataset_repositories.py
@@ -8,7 +8,7 @@ from dataall.core.environment.services.environment_service import EnvironmentSer
 from dataall.core.organizations.db.organization_repositories import Organization
 from dataall.base.db import paginate
 from dataall.base.db.exceptions import ObjectNotFound
-from dataall.modules.datasets_base.db.enums import ConfidentialityClassification, Language
+from dataall.modules.datasets_base.constants.enums import ConfidentialityClassification, Language
 from dataall.core.environment.services.environment_resource_manager import EnvironmentResource
 from dataall.modules.datasets_base.db.dataset_models import DatasetTable, Dataset
 from dataall.base.utils.naming_convention import (
@@ -177,7 +177,7 @@ class DatasetRepository(EnvironmentResource):
         ).to_dict()
 
     @staticmethod
-    def update_dataset_activity(session, dataset, username) :
+    def update_dataset_activity(session, dataset, username):
         activity = Activity(
             action='dataset:update',
             label='dataset:update',
